Improved SelectionScene visuals, Can now change airport

+++ src/main/java/uk/ac/soton/comp2211/airport/Runway.java
@@ -55,7 +55,7 @@ public class Runway {
         this.width = width;
     }
 
-    public int getBearing() {
+    public Double getBearing() {
         return bearing;
     }
